Working eureka discovery with config server + SSL
