+++ Eureka-Discovery/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,1 @@
+org.springframework.cloud.bootstrap.BootstrapConfiguration = com.example.eurekadiscovery.SslConfig
