+++ Eureka-Discovery/src/main/resources/application.yml
@@ -12,7 +12,7 @@ server:
 
 spring:
   application:
-    name:  Discovery
+    name:  discovery
 
 eureka:
   instance:
