+++ Eureka-Discovery/src/main/resources/bootstrap.yml
@@ -1,0 +1,10 @@
+spring:
+  cloud:
+    config:
+      uri: https://localhost:8889
+      failFast: true
+      retry:
+       max-attempts: 40
+  application:
+    name: discovery
+
